changed ID to a long instead of an int
git-svn-id: svn://svn.code.sf.net/p/ponykart/code/trunk@268 ba97e771-f6de-4074-9afc-763755853dd4

+++ Ponykart/Actors/Components/ModelComponent.cs
@@ -10,7 +10,7 @@ namespace Ponykart.Actors {
 	public class ModelComponent : LDisposable {
 		public SceneNode Node { get; protected set; }
 		public Entity Entity { get; protected set; }
-		public int ID { get; protected set; }
+		public long ID { get; protected set; }
 		public string Name { get; protected set; }
 
 		/// <summary>
